Create charts-original dir for non git packages
Adds a simple `mkdir -p ${f}/charts-original` to ensure the directory is
created before we try to cur and tar into it. The directory gets rm -rf
anyways at the end of the script.

coredns package failed to curl and tar due to failing to chdir into the
charts-original directory.

```
tar: could not chdir to 'packages/coredns/charts-original'
```